feat: ecs system almost updated

+++ Engine/Source/Assets/Mesh.cpp
@@ -20,7 +20,7 @@ namespace Engine
 
 	namespace
 	{
-		const BufferLayout c_defaultMeshLayoutInternal = 
+		const BufferLayout c_defaultMeshLayoutInternal =
 		{
 			BufferLayoutParameterSet::Position,
 			BufferLayoutParameterSet::Normal,
@@ -71,19 +71,19 @@ namespace Engine
 
 	void Mesh::SetIndices(const std::vector<uint32_t>& indices)
 	{
-		SetIndices(&indices[0], indices.size());
+		SetIndices(&indices[0], (uint32_t)indices.size());
 	}
 
 	void Mesh::SetIndices(const uint32_t* indices, uint32_t indexCount)
 	{
 		m_indices = new std::uint32_t[indexCount];
-		std::memcpy(m_indices, indices, 
+		std::memcpy(m_indices, indices,
 			sizeof(std::uint32_t) * indexCount);
-		
+
 		// Creates the index buffer.
 		std::shared_ptr<IndexBuffer> indexBuffer;
 		IndexBuffer::Create(indexBuffer, indices, indexCount, sizeof(uint32_t));
-		
+
 		if (m_vertexArray != nullptr)
 		{
 			m_vertexArray->SetIndexBuffer(indexBuffer);
@@ -106,7 +106,7 @@ namespace Engine
 				m_uvs[i].SetVertexCount(vertexCount);
 				m_uvs[i].SetBufferLayout({
 					{ "UV", BufferLayoutSemanticType::Type_TexCoord, BufferLayoutType::Float32, 2, i }
-				});
+					});
 			}
 			m_positions.SetVertexCount(vertexCount);
 			m_positions.SetBufferLayout(BufferLayoutParameterSet::Position);
@@ -130,7 +130,7 @@ namespace Engine
 	void Mesh::RefreshVertexArray()
 	{
 		m_vertexArray->ClearVertexBuffers();
-	
+
 		if (m_vertexCount > 0)
 		{
 			for (const auto& parameters : m_bufferLayout.parameters)
@@ -235,7 +235,7 @@ namespace Engine
 		RefreshVertexArray();
 	}
 
-	
+
 	void Mesh::SetUVs(uint32_t index, const std::vector<MathLib::Vector2>& uvs)
 	{
 		SetUVs(index, &uvs[0], uvs.size());
@@ -251,7 +251,7 @@ namespace Engine
 		const auto& buffer = uvIn.GetVertexBuffer();
 		buffer->SetBufferLayoutParameters({
 			 { "UV", BufferLayoutSemanticType::Type_TexCoord, BufferLayoutType::Float32, 2, index }
-		});
+			});
 		RefreshVertexArray();
 	}
 
@@ -268,7 +268,7 @@ namespace Engine
 		const auto& buffer = m_normals.GetVertexBuffer();
 		buffer->SetBufferLayoutParameters({
 			BufferLayoutParam::Normal0
-		});
+			});
 	}
 
 	void Mesh::RecalculateNormals()
